requirements.txt
behave
requests
Pillow
Faker